move public params
