+++ examples/minroot.rs
@@ -7,11 +7,12 @@ use bellpepper_core::{num::AllocatedNum, ConstraintSystem, SynthesisError};
 use ff::PrimeField;
 use flate2::{write::ZlibEncoder, Compression};
 use nova_snark::{
+  parameters::PublicParams,
   traits::{
-    circuit::{StepCircuit, TrivialCircuit},
+    circuit::{StepCircuit, TrivialTestCircuit},
     Group,
   },
-  CompressedSNARK, PublicParams, RecursiveSNARK,
+  CompressedSNARK, RecursiveSNARK,
 };
 use num_bigint::BigUint;
 use std::time::Instant;
@@ -84,9 +85,10 @@ where
   fn synthesize<CS: ConstraintSystem<F>>(
     &self,
     cs: &mut CS,
+    _pc: Option<&AllocatedNum<F>>,
     z: &[AllocatedNum<F>],
-  ) -> Result<Vec<AllocatedNum<F>>, SynthesisError> {
-    let mut z_out: Result<Vec<AllocatedNum<F>>, SynthesisError> =
+  ) -> Result<(Option<AllocatedNum<F>>, Vec<AllocatedNum<F>>), SynthesisError> {
+    let mut z_out: Result<(Option<AllocatedNum<F>>, Vec<AllocatedNum<F>>), SynthesisError> =
       Err(SynthesisError::AssignmentMissing);
 
     // use the provided inputs
@@ -119,7 +121,7 @@ where
       );
 
       if i == self.seq.len() - 1 {
-        z_out = Ok(vec![x_i_plus_1.clone(), x_i.clone()]);
+        z_out = Ok((None, vec![x_i_plus_1.clone(), x_i.clone()]));
       }
 
       // update x_i and y_i for the next iteration
@@ -157,7 +159,7 @@ fn main() {
       ],
     };
 
-    let circuit_secondary = TrivialCircuit::default();
+    let circuit_secondary = TrivialTestCircuit::default();
 
     println!("Proving {num_iters_per_step} iterations of MinRoot per step");
 
@@ -168,8 +170,8 @@ fn main() {
       G1,
       G2,
       MinRootCircuit<<G1 as Group>::Scalar>,
-      TrivialCircuit<<G2 as Group>::Scalar>,
-    >::new(&circuit_primary, &circuit_secondary, None, None);
+      TrivialTestCircuit<<G2 as Group>::Scalar>,
+    >::new_nova(&circuit_primary, &circuit_secondary, None, None);
     println!("PublicParams::setup, took {:?} ", start.elapsed());
 
     println!(
@@ -212,7 +214,7 @@ fn main() {
     let z0_secondary = vec![<G2 as Group>::Scalar::zero()];
 
     type C1 = MinRootCircuit<<G1 as Group>::Scalar>;
-    type C2 = TrivialCircuit<<G2 as Group>::Scalar>;
+    type C2 = TrivialTestCircuit<<G2 as Group>::Scalar>;
     // produce a recursive SNARK
     println!("Generating a RecursiveSNARK...");
     let mut recursive_snark: RecursiveSNARK<G1, G2, C1, C2> = RecursiveSNARK::<G1, G2, C1, C2>::new(
